add logpollerutil
