+++ core/services/ocr2/plugins/ccip/commit_plugin.go
@@ -17,6 +17,7 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/contractutil"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/hashlib"
+	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/logpollerutil"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/oraclelib"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/pricegetter"
 	"github.com/smartcontractkit/chainlink/v2/core/utils"
@@ -269,7 +270,7 @@ func unregisterCommitPluginFilters(ctx context.Context, sourceLP, destLP logpoll
 		return err
 	}
 
-	if err := unregisterLpFilters(
+	if err := logpollerutil.UnregisterLpFilters(
 		sourceLP,
 		getCommitPluginSourceLpFilters(staticCfg.OnRamp),
 		qopts...,
@@ -277,7 +278,7 @@ func unregisterCommitPluginFilters(ctx context.Context, sourceLP, destLP logpoll
 		return err
 	}
 
-	return unregisterLpFilters(
+	return logpollerutil.UnregisterLpFilters(
 		destLP,
 		getCommitPluginDestLpFilters(dynamicCfg.PriceRegistry, offRamp),
 		qopts...,
