+++ core/services/ocr2/plugins/ccip/plugins_common.go
@@ -5,18 +5,15 @@ import (
 	"math/big"
 	"sort"
 
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/pkg/errors"
 	"golang.org/x/exp/constraints"
 
-	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/logpoller"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/commit_store"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_offramp"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_onramp"
 	"github.com/smartcontractkit/chainlink/v2/core/logger"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/hashlib"
-	"github.com/smartcontractkit/chainlink/v2/core/services/pg"
 	"github.com/smartcontractkit/chainlink/v2/core/utils"
 )
 
@@ -74,67 +71,6 @@ func leavesFromIntervals(
 	}
 
 	return leaves, nil
-}
-
-func filterContainsZeroAddress(addrs []common.Address) bool {
-	for _, addr := range addrs {
-		if addr == utils.ZeroAddress {
-			return true
-		}
-	}
-	return false
-}
-
-func registerLpFilters(lp logpoller.LogPoller, filters []logpoller.Filter, qopts ...pg.QOpt) error {
-	for _, lpFilter := range filters {
-		if filterContainsZeroAddress(lpFilter.Addresses) {
-			continue
-		}
-		if err := lp.RegisterFilter(lpFilter, qopts...); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func unregisterLpFilters(lp logpoller.LogPoller, filters []logpoller.Filter, qopts ...pg.QOpt) error {
-	for _, lpFilter := range filters {
-		if filterContainsZeroAddress(lpFilter.Addresses) {
-			continue
-		}
-		if err := lp.UnregisterFilter(lpFilter.Name, qopts...); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func containsFilter(filters []logpoller.Filter, f logpoller.Filter) bool {
-	for _, existing := range filters {
-		if existing.Name == f.Name {
-			return true
-		}
-	}
-	return false
-}
-
-func filtersDiff(filtersBefore, filtersNow []logpoller.Filter) (created, deleted []logpoller.Filter) {
-	created = make([]logpoller.Filter, 0, len(filtersNow))
-	deleted = make([]logpoller.Filter, 0, len(filtersBefore))
-
-	for _, f := range filtersNow {
-		if !containsFilter(filtersBefore, f) {
-			created = append(created, f)
-		}
-	}
-
-	for _, f := range filtersBefore {
-		if !containsFilter(filtersNow, f) {
-			deleted = append(deleted, f)
-		}
-	}
-
-	return created, deleted
 }
 
 func max[T constraints.Ordered](first T, rest ...T) T {
